# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * InterviewAce AI - Career preparation platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  clerkId: string;
  email: string;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  role: string;
  /** @nullable */
  targetRole?: string | null;
  /** @nullable */
  experienceLevel?: string | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  name?: string;
  targetRole?: string;
  experienceLevel?: string;
}

export interface DashboardStats {
  /** @nullable */
  resumeScore: number | null;
  interviewCount: number;
  /** @nullable */
  avgTechnicalScore: number | null;
  /** @nullable */
  avgCommunicationScore: number | null;
  /** @nullable */
  avgCodingScore: number | null;
  weakTopics: string[];
  totalSubmissions: number;
}

export interface ActivityItem {
  id: number;
  type: string;
  title: string;
  description: string;
  /** @nullable */
  score?: number | null;
  createdAt: string;
}

/**
 * JSON blob with resume sections
 */
export type ResumeContent = { [key: string]: unknown };

export interface Resume {
  id: number;
  userId: number;
  title: string;
  template: string;
  /** JSON blob with resume sections */
  content: ResumeContent;
  /** @nullable */
  atsScore: number | null;
  createdAt: string;
  updatedAt: string;
}

export type ResumeInputContent = { [key: string]: unknown };

export interface ResumeInput {
  title: string;
  template?: string;
  content?: ResumeInputContent;
}

export type ResumeUpdateContent = { [key: string]: unknown };

export interface ResumeUpdate {
  title?: string;
  template?: string;
  content?: ResumeUpdateContent;
  atsScore?: number;
}

export interface ResumeAnalysis {
  score: number;
  atsCompatibility: number;
  missingKeywords: string[];
  grammarSuggestions: string[];
  recommendations: string[];
  strengths?: string[];
}

export interface InterviewSession {
  id: number;
  userId: number;
  category: string;
  difficulty: string;
  status: string;
  /** @nullable */
  technicalScore: number | null;
  /** @nullable */
  communicationScore: number | null;
  /** @nullable */
  confidenceScore: number | null;
  questionsAsked: number;
  answersGiven: number;
  answeredQuestionIds?: number[];
  createdAt: string;
}

export type InterviewInputCategory = typeof InterviewInputCategory[keyof typeof InterviewInputCategory];


export const InterviewInputCategory = {
  java: 'java',
  python: 'python',
  mern: 'mern',
  fullstack: 'fullstack',
  data_analyst: 'data_analyst',
  data_science: 'data_science',
  ai_ml: 'ai_ml',
  hr: 'hr',
} as const;

export type InterviewInputDifficulty = typeof InterviewInputDifficulty[keyof typeof InterviewInputDifficulty];


export const InterviewInputDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface InterviewInput {
  category: InterviewInputCategory;
  difficulty: InterviewInputDifficulty;
}

export interface AnswerInput {
  questionId: number;
  answer: string;
  /** Optional personalized wording shown to the candidate for this question */
  questionText?: string;
}

export interface AnswerEvaluation {
  id: number;
  interviewId: number;
  questionId: number;
  answer: string;
  technicalScore: number;
  communicationScore: number;
  confidenceScore: number;
  correctAnswer: string;
  feedback: string;
  improvements?: string[];
}

export interface Question {
  id: number;
  category: string;
  difficulty: string;
  text: string;
  type: string;
  /** @nullable */
  sampleAnswer?: string | null;
  tags?: string[];
}

export type CodingProblemStarterCode = { [key: string]: unknown };

export type CodingProblemExamplesItem = { [key: string]: unknown };

export interface CodingProblem {
  id: number;
  title: string;
  description: string;
  difficulty: string;
  languages: string[];
  starterCode: CodingProblemStarterCode;
  examples: CodingProblemExamplesItem[];
  /** @nullable */
  constraints?: string | null;
  timeLimit?: number;
}

export interface CodeSubmissionInput {
  problemId: number;
  language: string;
  code: string;
}

export interface CodingSubmission {
  id: number;
  userId: number;
  problemId: number;
  language: string;
  code: string;
  score: number;
  passed: boolean;
  testsPassed?: number;
  testsTotal?: number;
  aiReview: string;
  complexity: string;
  suggestions?: string[];
  createdAt: string;
}

export interface RoadmapStep {
  week: number;
  topic: string;
  goal: string;
  resources: string[];
}

export interface VideoResource {
  title: string;
  url: string;
  topic: string;
  duration: string;
}

export interface LearningRoadmap {
  weakTopics: string[];
  strongTopics: string[];
  plan: RoadmapStep[];
  recommendedVideos: VideoResource[];
}

export interface LearningTopic {
  topic: string;
  strength: string;
  questionsAttempted: number;
  avgScore: number;
}

export interface AdminUser {
  id: number;
  clerkId: string;
  email: string;
  name: string;
  role: string;
  interviewCount: number;
  resumeCount: number;
  createdAt: string;
}

export type AdminAnalyticsUserGrowthItem = { [key: string]: unknown };

export interface AdminAnalytics {
  totalUsers: number;
  totalInterviews: number;
  totalResumes: number;
  totalSubmissions: number;
  avgInterviewScore: number;
  userGrowth: AdminAnalyticsUserGrowthItem[];
}

export type ListQuestionsParams = {
category?: string;
difficulty?: string;
limit?: number;
};

export type ListCodingProblemsParams = {
difficulty?: string;
language?: string;
};