# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: InterviewAce AI - Career preparation platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /interviews:
    get:
      operationId: listInterviews
      tags: [interviews]
      summary: List user interview sessions
      responses:
        "200":
          description: Interview sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InterviewSession"
    post:
      operationId: createInterview
      tags: [interviews]
      summary: Start a new mock interview session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterviewInput"
      responses:
        "201":
          description: Created interview session with first question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
  /interviews/{id}:
    get:
      operationId: getInterview
      tags: [interviews]
      summary: Get interview session details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"
        "404":
          description: Not found
    delete:
      operationId: deleteInterview
      tags: [interviews]
      summary: Delete an interview session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /interviews/{id}/answers:
    post:
      operationId: submitAnswer
      tags: [interviews]
      summary: Submit an answer and get AI evaluation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "201":
          description: Answer evaluation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerEvaluation"
  /interviews/{id}/abandon:
    post:
      operationId: abandonInterview
      tags: [interviews]
      summary: End an interview when the candidate leaves the interview tab
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Interview ended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewSession"

  # ─── Questions ────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    InterviewSession:
      type: object
      required: [id, userId, category, difficulty, status, technicalScore, communicationScore, confidenceScore, questionsAsked, answersGiven, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        category:
          type: string
        difficulty:
          type: string
        status:
          type: string
        technicalScore:
          type: ["number", "null"]
        communicationScore:
          type: ["number", "null"]
        confidenceScore:
          type: ["number", "null"]
        questionsAsked:
          type: integer
        answersGiven:
          type: integer
        answeredQuestionIds:
          type: array
          items:
            type: integer
        createdAt:
          type: string
    InterviewInput:
      type: object
      required: [category, difficulty]
      properties:
        category:
          type: string
          enum: [java, python, mern, fullstack, data_analyst, data_science, ai_ml, hr]
        difficulty:
          type: string
          enum: [beginner, intermediate, advanced]
    AnswerInput:
      type: object
      required: [questionId, answer]
      properties:
        questionId:
          type: integer
        answer:
          type: string
        questionText:
          type: string
          description: Optional personalized wording shown to the candidate for this question
    AnswerEvaluation:
      type: object
      required: [id, interviewId, questionId, answer, technicalScore, communicationScore, confidenceScore, correctAnswer, feedback]
      properties:
        id:
          type: integer
        interviewId:
          type: integer
        questionId:
          type: integer
        answer:
          type: string
        technicalScore:
          type: integer
        communicationScore:
          type: integer
        confidenceScore:
          type: integer
        correctAnswer:
          type: string
        feedback:
          type: string
        improvements:
          type: array
          items:
            type: string

    # ─── Questions ─────────────────────────────────────────────────────────
...