# lib/api-zod/src/generated/types/answerInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * InterviewAce AI - Career preparation platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AnswerInput {
  questionId: number;
  answer: string;
  /** Optional personalized wording shown to the candidate for this question */
  questionText?: string;
}